added

+++ src/main/java/skill/project/dto/TagDto.java
@@ -3,9 +3,11 @@ package skill.project.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class TagDto {
   private String name;
-  private Double weight;
+  private BigDecimal weight;
 }
