+++ src/main/java/skill/project/dto/response/CalendarResponse.java
@@ -1,10 +1,9 @@
 package skill.project.dto.response;
 
 import lombok.AllArgsConstructor;
-import lombok.Data;
 import lombok.Getter;
 
-import java.util.HashMap;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
@@ -12,5 +11,5 @@ import java.util.Map;
 @AllArgsConstructor
 public class CalendarResponse {
   List<Integer> years;
-  Map<String, Integer> posts;
+  Map<String, BigInteger> posts;
 }
